Stabilize MAX77958 opcode role handling
Serialize MAX77958 opcode dispatch with an in-flight guard so interrupt snapshots cannot start overlapping opcode commands.

Add bounded DR_SWAP retry handling, clear delayed recheck pending flags when alarm delivery fails, and move the external MAX77958 I2C1 lab test out of robot.c into its own module.
