+++ include/external_max77958_test.h
@@ -1,0 +1,6 @@
+#ifndef EXTERNAL_MAX77958_TEST_H_
+#define EXTERNAL_MAX77958_TEST_H_
+
+void external_max77958_i2c1_test(void);
+
+#endif
